- implemented getting list of notes by notebook - implemented note deletion - refactored and removed a few models not being used

+++ src/main/java/com/dougb/dougbook/models/NoteModel.java
@@ -21,7 +21,7 @@ public class NoteModel {
     }
 
 
-    @Id @GeneratedValue private int id;
+    @Id @GeneratedValue private long id;
     @Getter @Setter @Column(nullable = false)  private String title;
     @ElementCollection @Column(nullable = false) private List<String> tag = new ArrayList<>();
     @Getter @Setter @Column(nullable = false) private String body;
